feat: FAQ 카드에 애니메이션 추가
